Use remove function as module
